+++ dio/lib/sentry_dio.dart
@@ -1,0 +1,3 @@
+library sentry_dio;
+
+export 'src/sentry_client_adapter.dart';