+++ dio/lib/src/breadcrumb_client_adapter.dart
@@ -1,0 +1,70 @@
+// ignore_for_file: strict_raw_type
+
+import 'dart:typed_data';
+
+import 'package:dio/adapter.dart';
+import 'package:dio/dio.dart';
+import 'package:sentry/sentry.dart';
+
+/// A [Dio](https://pub.dev/packages/dio)-package compatible HTTP client adapter
+/// which records requests as breadcrumbs.
+///
+/// Remarks:
+/// If this client is used as a wrapper, a call to close also closes the
+/// given client.
+class BreadcrumbClientAdapter extends HttpClientAdapter {
+  // ignore: public_member_api_docs
+  BreadcrumbClientAdapter({HttpClientAdapter? client, Hub? hub})
+      : _hub = hub ?? HubAdapter(),
+        _client = client ?? DefaultHttpClientAdapter();
+
+  final HttpClientAdapter _client;
+  final Hub _hub;
+
+  @override
+  Future<ResponseBody> fetch(
+    RequestOptions options,
+    Stream<Uint8List>? requestStream,
+    Future? cancelFuture,
+  ) async {
+    // See https://develop.sentry.dev/sdk/event-payloads/breadcrumbs/
+
+    var requestHadException = false;
+    int? statusCode;
+    String? reason;
+    int? responseBodySize;
+
+    final stopwatch = Stopwatch();
+    stopwatch.start();
+
+    try {
+      final response =
+          await _client.fetch(options, requestStream, cancelFuture);
+
+      statusCode = response.statusCode;
+      reason = response.statusMessage;
+
+      return response;
+    } catch (_) {
+      requestHadException = true;
+      rethrow;
+    } finally {
+      stopwatch.stop();
+
+      final breadcrumb = Breadcrumb.http(
+        level: requestHadException ? SentryLevel.error : SentryLevel.info,
+        url: options.uri,
+        method: options.method,
+        statusCode: statusCode,
+        reason: reason,
+        requestDuration: stopwatch.elapsed,
+        responseBodySize: responseBodySize,
+      );
+
+      _hub.addBreadcrumb(breadcrumb);
+    }
+  }
+
+  @override
+  void close({bool force = false}) => _client.close(force: force);
+}