+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+# Unreleased
+
 # 6.3.0-alpha.1
 
 * Feat: Automatically create transactions when navigating between screens (#643)