+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* Enha: Replace flutter default root name '/' with 'root' (#678)
+
 # 6.3.0-alpha.1
 
 * Feat: Automatically create transactions when navigating between screens (#643)