+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 * Fix: Use 'navigation' instead of 'ui.load' for auto transaction operation (#675)
+* Add `SentryAssetBundle` for automatic spans for asset loading (#685)
 
 # 6.3.0-alpha.1