+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased
 
+* Fix: Use 'navigation' instead of 'ui.load' for auto transaction operation (#675)
+* Fix: Use correct data/extras type in tracer (#693)
+
 # 6.3.0-alpha.1
 
 * Feat: Automatically create transactions when navigating between screens (#643)