+++ CHANGELOG.md
@@ -2,11 +2,11 @@
 
 * Fix: Use 'navigation' instead of 'ui.load' for auto transaction operation (#675)
 * Fix: Use correct data/extras type in tracer (#693)
+* Fix: Do not throw when Throwable type is not supported for associating errors to a transaction (#692)
 
 # 6.3.0-alpha.1
 
 * Feat: Automatically create transactions when navigating between screens (#643)
-* Fix: Do not throw when Throwable type is not supported for associating errors to a transaction (#692)
 
 # 6.2.2