+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Enha: Replace flutter default root name '/' with 'root' (#678)
 * Fix: Use 'navigation' instead of 'ui.load' for auto transaction operation (#675)
 * Fix: Use correct data/extras type in tracer (#693)
 * Fix: Do not throw when Throwable type is not supported for associating errors to a transaction (#692)