+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 6.3.0-beta.1
 
 * Enha: Replace flutter default root name '/' with 'root' (#678)
 * Fix: Use 'navigation' instead of 'ui.load' for auto transaction operation (#675)